Add follower log hole support.    - When follower log exist hole, leader need to fill the hole first

+++ RaftLib/RaftNoteBook.cs
@@ -124,9 +124,9 @@ namespace PineHillRSL.Raft.Notebook
 
         private UInt64 _maxDecreeNo = 0;
         private AppendPosition _lastAppendPosition;
-        private Int64 _maxCommittedLogIndex = 0;
+        private UInt64 _maxCommittedLogIndex = 0;
         private UInt64 _maxCommittedLogTerm = 0;
-        private Int64 _maxLogIndex = 0;
+        private UInt64 _maxLogIndex = 0;
         private UInt64 _maxTerm = 0;
         private CancellationTokenSource _cancel = new CancellationTokenSource();
 
@@ -194,14 +194,14 @@ namespace PineHillRSL.Raft.Notebook
                     LogPosition = it.Position,
                     Content = entityRecord.DecreeContent
                 });
-                if ((Int64)entityRecord.CommittedLogIndex > _maxCommittedLogIndex)
+                if (entityRecord.CommittedLogIndex > _maxCommittedLogIndex)
                 {
-                    _maxCommittedLogIndex = (Int64)entityRecord.CommittedLogIndex;
+                    _maxCommittedLogIndex = entityRecord.CommittedLogIndex;
                 }
 
-                if ((Int64)entityRecord.LogIndex > _maxLogIndex)
+                if (entityRecord.LogIndex > _maxLogIndex)
                 {
-                    _maxLogIndex = (Int64)entityRecord.LogIndex;
+                    _maxLogIndex = entityRecord.LogIndex;
                 }
 
                 if (entityRecord.Term > _maxTerm)
@@ -252,13 +252,13 @@ namespace PineHillRSL.Raft.Notebook
                     LogPosition = position
                 });
 
-                if ((Int64)committedLogIndex > _maxCommittedLogIndex)
+                if (committedLogIndex > _maxCommittedLogIndex)
                 {
-                    _maxCommittedLogIndex = (Int64)committedLogIndex;
+                    _maxCommittedLogIndex = committedLogIndex;
                 }
-                if ((Int64)logIndex > _maxLogIndex)
+                if (logIndex > _maxLogIndex)
                 {
-                    _maxLogIndex = (Int64)logIndex;
+                    _maxLogIndex = logIndex;
                 }
 
                 if (term > _maxTerm)
@@ -331,10 +331,10 @@ namespace PineHillRSL.Raft.Notebook
             return _logEntities;
         }
 
-        public Int64 CommittedLogIndex => _maxCommittedLogIndex;
+        public UInt64 CommittedLogIndex => _maxCommittedLogIndex;
         public UInt64 LastCommittedLogTerm => _maxCommittedLogTerm;
 
-        public Int64 MaxLogIndex => _maxLogIndex;
+        public UInt64 MaxLogIndex => _maxLogIndex;
 
         public UInt64 MaxTerm => _maxTerm;
     }
